% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{daily2monthly}
\alias{daily2monthly}
\title{Convert daily data to monthly (zoo)----}
\usage{
daily2monthly(x, FUN, na.rm = TRUE, ...)
}
\arguments{
\item{x}{Input zoo variable.}

\item{FUN}{function for aggregation}
}
\description{
Convert daily data to monthly (zoo)----
}
\examples{
daily2monthly( FUN=mean)

}
\keyword{aggregation}
