% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dWaSSI_functions.R
\name{distHydroSim}
\alias{distHydroSim}
\title{Distribution wrap of dWaSSI-C}
\usage{
distHydroSim(sim.dates, par.sacsma = NULL, par.petHamon = NULL,
  par.snow17 = NULL, par.routLah = NULL, hru.info = NULL,
  hru.elevband = NULL, clim.prcp = NULL, clim.tavg = NULL,
  hru.lai = NULL, hru.lc.lai = NULL, huc.lc.ratio = NULL,
  snow.flag = 0, progress.bar = TRUE, month = F)
}
\arguments{
\item{sim.dates}{list of all dates}
}
\value{
outputs potential evapotranspiration (mm day-1)
}
\description{
FUNCTION_DESCRIPTION
}
\details{
For details see Haith and Shoemaker (1987)
}
\examples{
\dontrun{
distHydroSim(str.date,
             hru.par = hru_par, hru.info = hru_info, hru.elevband = NULL,
             clim.prcp = stcroix$prcp.grid, clim.tavg = stcroix$tavg.grid,
             snow.flag = 0, progress.bar = TRUE)
}
}
