% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{f_2nc}
\alias{f_2nc}
\title{Write a file to netcdf file}
\usage{
f_2nc(filename, ncfname, varname, start_date = NA, scale = "1 year",
  attrs = NA, fillvalue = -99, lname = "", varunit = "", plot = T)
}
\arguments{
\item{filename}{The input nc file}

\item{ncfname}{The output nc file}

\item{varname}{This is the var names for the data}

\item{start_date}{Optional. The start date for the input data ("1982-01-01")}

\item{scale}{Optional. This scale of the time series. ("1 year","1 month", "1 day")}

\item{attrs}{Optional. Extra attributes for the data. c("name"="value")}

\item{fillvalue}{Optional. The default fill value for the missing data.}

\item{lname}{The long name of the variable}

\item{varunit}{The unit of the variable}

\item{plot}{Logical. Whether plot the output nc"}
}
\description{
Write a file to netcdf file
}
\examples{
path and file name, set dname
# add global attributes
attr_global<-c("Author"="Ning Liu",
              "Email"="N.LIU@murdoch.edu.au",
              "Institution"="Murdoch University",
              "References"="Reference")
f_2nc(filename="/mnt/Ning/GPP_Tr_anomaly_ann_Fixed_CO2_82_14.nc",
ncfname="/mnt/Ning/tt.nc",
varname = "GPP_Tr",
varunit= "g C kg-1 H2O"
lname = "Anomaly of annual GPP-Tr with Fixed CO2 from CALBE model",
start_date = "1982-01-01",
scale = "1 month")
}
\keyword{NetCDF}
\keyword{write}
