% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{f_2raster}
\alias{f_2raster}
\title{Transfering matrix/array to raster/brick
This function allows you to transfer a matrix/array to a raster/brick.}
\usage{
f_2raster(data, infonc = NA)
}
\arguments{
\item{data}{A matrix or array object.}

\item{infonc}{A filename of a raster object for getting the raster extent info.}
}
\description{
Transfering matrix/array to raster/brick
This function allows you to transfer a matrix/array to a raster/brick.
}
\examples{
rc<-f_2raster(darray,infonc="/Dataset/backup/CABLE/ET_ann_82_14.nc")

}
\keyword{cats}
