% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{f_cor}
\alias{f_cor}
\title{Calculate the correlation coefficient between two vectors}
\usage{
f_cor(da, method = "spearman")
}
\arguments{
\item{da}{The input is a vector, which is combination of c(x,y)}

\item{method}{One of "pearson", "kendall", or "spearman", can be abbreviated.}
}
\description{
Calculate the correlation coefficient between two vectors
}
\examples{
x<-c(1:10);y<-c(1:10)
damerge<-c(x,y)
f_cor(da=damerge,method="spearman")

}
\keyword{correlationship}
