% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{f_lib_check}
\alias{f_lib_check}
\title{This function allows you to check whether the required library has been installed, otherwise it will be installed and load.}
\usage{
f_lib_check(libs)
}
\arguments{
\item{libs}{A character vector of names of required libraries.}
}
\description{
This function allows you to check whether the required library has been installed, otherwise it will be installed and load.
}
\examples{
libs<-c("ggplot2","caTools")
f_lib_check(libs)
}
\keyword{libraries}
