% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{f_line_plot}
\alias{f_line_plot}
\title{Plot annual mean line}
\usage{
f_line_plot(name1)
}
\arguments{
\item{...}{ggplot plots.}

\item{cols}{Number of columns}
}
\description{
Plot annual mean line
}
\examples{
multiplot(p1,p2,p3,p4,cols=2)

}
