% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{f_mon2annual_array}
\alias{f_mon2annual_array}
\title{Convert a monthly array to a annual array}
\usage{
f_mon2annual_array(da, fun = "mean")
}
\arguments{
\item{da}{The input monthly array}

\item{fun}{The function for aggregation: "mean", "sum"}
}
\description{
Convert a monthly array to a annual array
}
\examples{
da<-array(c(1:288),c(4,3,24))
f_mon2annual_array(da=da,fun="mean")
}
\keyword{trend}
