% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{f_Parallel_set}
\alias{f_Parallel_set}
\title{Setup up parallel using FORK}
\usage{
f_Parallel_set(name = "zeus", ncores = NA)
}
\arguments{
\item{name}{A filename for storing parallel log.}

\item{ncores}{How many cores will be used for parallelization}
}
\description{
Setup up parallel using FORK
}
\examples{
f_Parallel_set(name="zeus",ncores=10)

}
\keyword{Parallel}
