% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{f_paste}
\alias{f_paste}
\title{Paste by value one to one for two vectors, matrixes or arrays}
\usage{
f_paste(x, y, sep = "")
}
\arguments{
\item{x}{The first object, which can be vector, matrix or array.}

\item{y}{The second object, which can be vector, matrix or array.}

\item{sep}{The separate letter}
}
\description{
Paste by value one to one for two vectors, matrixes or arrays
}
\examples{
x<-c(1,2,3)
y<-c("A","B","C")
f_paste(x,y,sep="-")

}
\keyword{paste}
