% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{f_plot_sp}
\alias{f_plot_sp}
\title{Plot spatial data with overlap shpfile}
\usage{
f_plot_sp(da, filename, colstyle = "RdYlGn", pretty = T, margin = list(),
  shpname = NA, varnames = NA, cuts = NA, ranges = NA, width = 7,
  height = 7)
}
\arguments{
\item{da}{The input nc file.}

\item{filename}{The input polygon.}

\item{colstyle}{"RdYlGn",}

\item{pretty}{Logic}

\item{margin}{list or logic. This is the Statistic information for the data showed on the margin}

\item{shpname}{It can be any sting if a shp will be ploted}

\item{varnames}{This is the var names for the plot variable}

\item{cuts}{This is how many segments of the plot}

\item{ranges}{This is the min and max of the data}

\item{width}{This is the width of the pdf in "inch"}

\item{height}{This is the height of the pdf in "inch"}
}
\description{
Plot spatial data with overlap shpfile
}
\examples{
# the shpfile has to be loaded before the plot
library(maptools)
shp<-readShapeLines("data/shp/AU_STATES.shp")
f_plot_sp(da = WTD,
       filename = "/Dataset/www/images/WUE_AU/tt1.pdf",
       width=7,height=7,
       cuts = 10,
       ranges = c(3,100),
       shpname = "shp")
}
\keyword{plot}
\keyword{spatial}
