% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{f_sta_shp_nc}
\alias{f_sta_shp_nc}
\title{Zonal raster/brick based on a shapefile}
\usage{
f_sta_shp_nc(ncfilename, basin, fun = "mean", varname, zonal_field, start,
  scale = "month", df = T)
}
\arguments{
\item{ncfilename}{A filename for a "raster*" type object.}

\item{basin}{A ploygon object.}

\item{fun}{function for doing zonal (fun="mean"/"sum")}

\item{varname}{define the variable name}

\item{zonal_field}{select the field from shapefile file for naming the result}

\item{start}{the start year for the time series of the input raster}

\item{scale}{the time step the the input raster}
}
\description{
Zonal raster/brick based on a shapefile
}
\examples{
sta_shp<-f_sta_shp_nc(ncfilename="/Dataset/backup/CABLE/ET_ann_82_14.nc",
basin,fun="sum",varname="ET",zonal_field="Station",start=1982,scale="annual")

}
\keyword{zonal}
