% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{f_trend}
\alias{f_trend}
\title{Trend analysis using linear regression}
\usage{
f_trend(data)
}
\arguments{
\item{da}{The input vector}
}
\description{
Trend analysis using linear regression
}
\examples{
x<-c(1:10)
f_trend(data=x)
}
\keyword{trend}
