% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dWaSSI_functions.R
\name{f_WaSSI}
\alias{f_WaSSI}
\title{A f_WaSSI function}
\usage{
f_WaSSI(data_in, pars, soil_in, calibrate = NA, y_s = NA, y_e = NA,
  scale = "month")
}
\arguments{
\item{data_in}{this is the monthly dataframe data, which includes all main variabels (P, T, E, Q etc.)}

\item{pars}{A names vector inlcudes "ALT","LAT","LONG",and "VEG" infomation}

\item{soil_in}{A names vector inlcudes all soil parameters}

\item{calibrate}{Default =NA, defines wheter calibrate model in watershed scale}

\item{y_s, y_e}{The start and end year for simulation}

\item{scale}{Simulation scale c("month","ann")}
}
\description{
Calculate Q with SMA-SMC and snow model
}

