% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{f_zonal_shp_nc}
\alias{f_zonal_shp_nc}
\title{Zonal catergory raster file based on shp}
\usage{
f_zonal_shp_nc(ncfilename, shp, zonal_field, category = T)
}
\arguments{
\item{ncfilename}{The input nc file.}

\item{shp}{The input polygon.}

\item{zonal_field}{elect the field from shapefile file for naming the result}
}
\description{
Zonal catergory raster file based on shp
}
\examples{
zonal_shp<-f_zonal_shp_nc(ncfilename="/Dataset/backup/CABLE/ET_ann_82_14.nc",
basin,,zonal_field="Station")

}
\keyword{zonal}
