% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{hru_zonal}
\alias{hru_zonal}
\title{Zonal a variable for each Hru in dWaSSI-C}
\usage{
hru_zonal(classname, daname, shp, fun = "mean", field = NULL,
  plot = T)
}
\arguments{
\item{classname}{a raster of each hru it can be vegetation type or soil type}

\item{daname}{The input brick data that will be zonaled by hru}

\item{shp}{the zonal boundary}

\item{field}{the field of the shp boundary that will used for zonal}

\item{varname}{The name of the zonaled variable}
}
\description{
FUNCTION_DESCRIPTION
}
\details{
This is a function for zonal hru data
}
\examples{
\dontrun{
ha<-hru_zonal(classname = "inputs/Landcover/LUCC_Sun_IGBP.nc",
             daname = "inputs/LAI/LAI_BNU.nc",
            shp = Basins,
            field="Station")
}
}
