% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dWaSSI_functions.R
\name{hrurouting}
\alias{hrurouting}
\title{hucrouting}
\usage{
hrurouting(water, routpar, mc_cores = 1)
}
\arguments{
\item{water}{this is the original flow data for each huc}

\item{routpar}{flow direction between hucs}
}
\description{
Calculate accumulated flow for each huc, based on the flow direction
}
\examples{
hrurouting(water,outpar)
}
