% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{multiplot}
\alias{multiplot}
\title{multiplot for ploting multi row and cols ggplot function}
\usage{
multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\arguments{
\item{...}{ggplot plots.}

\item{cols}{Number of columns}
}
\description{
multiplot for ploting multi row and cols ggplot function
}
\examples{
multiplot(p1,p2,p3,p4,cols=2)

}
