% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{numberOfDays}
\alias{numberOfDays}
\title{This function allows you to get the number of days for a specific month.}
\usage{
numberOfDays(date)
}
\arguments{
\item{date}{A date object.}
}
\description{
This function allows you to get the number of days for a specific month.
}
\examples{
date<-as.Date("2001-01-01")
numberOfDays(date)
}
\keyword{cats}
