% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sacsma_mon.R
\name{sacSma_mon}
\alias{sacSma_mon}
\title{Sacremento Soil Moisture Accounting Model SAC-SMA}
\usage{
sacSma_mon(pet, prcp, par, ini.states = c(0, 0, 500, 500, 500, 0))
}
\arguments{
\item{pet}{potential evapotranspiration, in mm}

\item{prcp}{daily precipitation data}

\item{par}{model parameters (11 soil parameters)}

\item{ini.states}{initial parameters}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
revised based on sacsmaR package
}
\details{
DETAILS
}
\examples{
\dontrun{
sacSma_mon(pet, prcp,par)
}
}
