% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dWaSSI_functions.R
\name{snow_melt}
\alias{snow_melt}
\title{SNOW Model}
\usage{
snow_melt(ts.prcp, ts.temp, meltmax = 0.5, snowrange = c(-3, 1),
  snowpack = 0)
}
\arguments{
\item{ts.prcp}{numeric vector of precipitation time-series (mm)}

\item{ts.temp}{numeric vector of average temperatuer time-series (Deg C)}

\item{meltmax}{maximium ratio of snow melt (default: 0.5)}

\item{snowrange}{(-3,1) the temperature range between snow and rain}

\item{snowpack}{initial snowpack (default: 0)}
}
\value{
a dataframe of c("prcp","snowpack","snowmelt"), effective rainfall, snowpack and snowmelt
}
\description{
Peter's Snow model
}
\details{
This is based on Peter's Fortran code
}
\examples{
\dontrun{
mellt<-snow_melt(ts.prcp,ts.temp,meltmax=0.5,snowrange=c(-3,1),snowpack=0)
}
}
