% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dWaSSI_functions.R
\name{dWaSSIC}
\alias{dWaSSIC}
\title{Parallel wrap of WaSSI-C model}
\usage{
dWaSSIC(sim.dates, warmup = 3, mcores = 1, par.sacsma = NULL,
  par.petHamon = NULL, par.routing = NULL, hru.info = NULL,
  clim.prcp = NULL, clim.tavg = NULL, hru.lai = NULL,
  hru.lc.lai = NULL, huc.lc.ratio = NULL)
}
\arguments{
\item{sim.dates}{list of all dates}

\item{warmup}{years for warming up WaSSI-C model}

\item{mcores}{how many cores using for simulation}
}
\value{
outputs potential evapotranspiration (mm day-1)
}
\description{
FUNCTION_DESCRIPTION
}
\details{
For details see Haith and Shoemaker (1987)
}
\examples{
\dontrun{
distHydroSim(sim.dates = sim_dates,
             hru.par = hru_par, hru.info = hru_info, hru.elevband = NULL,
             clim.prcp = stcroix$prcp.grid, clim.tavg = stcroix$tavg.grid,
             snow.flag = 0, progress.bar = TRUE)
}
}
