% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numberOfDays.R
\name{numberOfDays}
\alias{numberOfDays}
\title{A numberOfDays Function}
\usage{
numberOfDays(date)
}
\arguments{
\item{date}{Do you love cats? Defaults to TRUE.}

\item{da}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
numberOfDays("2001-01-01")
this is ()
}
\keyword{cats}

